requirements.txt
annotated-types==0.6.0
certifi          ==  2023.11.17
charset-normalizer ==3.3.2
idna               ==3.6
instagrapi
numpy
pandas
pillow             ==10.2.0

pyarrow            ==15.0.0
pycryptodomex      ==3.20.0
pydantic           ==2.5.3
pydantic_core      ==2.14.6
PySocks            ==1.7.1
python-dateutil    ==2.8.2
pytz               ==2023.4
requests           ==2.31.0
six                ==1.16.0
typing_extensions  ==4.9.0
tzdata             ==2023.4
urllib3            ==2.2.0
wheel              ==0.41.2
